test: Add property tests for `uri` type (#7)
Add tests to ensure that URI parsing and serializing works as expected.
Since the URI type itself is imported from network-uri, we don't need to
make the testing very robust. Instead, we simply add sanity checks to
ensure that the megaparsec-based URI parser behaves as expected,
especially when there are stray leading or trailing characters.