build: List all modules for test suite (#183)
The singletons code we are adding will need to depend on the GHC
version. To implement this, we will have two modules for singletons
code, one for old GHCs and for new GHCs, add logic in package.yaml to
select the right module at build time.

We can only do this kind of conditional build with an explicit list of
the modules that we are building. Therefore, we switch from implicitly
to explicitly specifying the test suite modules in package.yaml.