refactor: Re-export unified `SChar` and `SSymbol`
When GHC < 9.6, there is no `SChar` or `SSymbol` available in `base`,
but they are available from `singletons-base`. We want to minimize the
dependency on `singletons-base`, so

* For GHC < 9.6, we use `SChar` and `SSymbol` from `singletons-base`.
* For GHC >= 9.6, we use `SChar` and `SSymbol` from `base`.

And we re-export these types in `VCard.Symbol.Private`. This allows us
to have a consistent source for `SChar` and `SSymbol` regardless of GHC
version.