ci: Stop running CI automatically on push to main
Currently there are two ways that the CI pipeline is triggered:

  1. Creating a pull request targeted at main will trigger the CI
     pipeline on the latest commit of the pull request.

  2. Pushing to main will trigger the CI pipeline on the latest commit
     of main.

I have configured the Codeberg repo so that all PRs merged to main are
fast-forwards. Therefore, after a merge, the latest commit of the PR
becomes the latest commit on main. So, there will already be a passing
CI pipeline on this latest commit via (1) above, so there is no need to
run a pipeline via (2).

Therefore, to avoid duplicate CI pipelines, this commit disables (2).
Occasionally, we push to main directly without a PR, and there will be
no pipeline automatically triggered for the latest commit on main. In
this case, we can still trigger a pipeline run manually.