refactor: Minor fix of `IsXNameUpperSymbol` (#225)
Fix the definition of `IsXNameUpperSymbol s` to check
`IsPrefixOf "X-" s` instead of `IsPrefixOfInsensitive "X-" s`.

This doesn't actually change the behavior or `IsXNameUpperSymbol`, since
there is another check to make sure that all the character of `s` are
upper case, but it does make the implementation a bit more clear and
morally correct.